+++ Shop.Application/Services/CTSanPhamService.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Shop.Application.Exeptions;
 using Shop.Application.IServices;
-using Shop.Application.ViewModels;
 using Shop.Data.Context;
 using Shop.Data.Models;
+using Shop.ViewModels.ViewModels;
 
 namespace Shop.Application.Services;
