+++ Shop.ViewModels/Shop.ViewModels.csproj
@@ -6,9 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Shop.Data\Shop.Data.csproj" />
-    <ProjectReference Include="..\Shop.ViewModels\Shop.ViewModels.csproj" />
-  </ItemGroup>
-
 </Project>
