+++ Shop.ViewModels/ViewModels/ChucVuVM.cs
@@ -1,9 +1,9 @@
-﻿namespace Shop.Application.ViewModels;
+﻿namespace Shop.ViewModels.ViewModels;
 
-public class KichCoVM
+public class ChucVuVM
 {
     public Guid Id { get; set; }
     public string Ma { get; set; }
-    public int SoSize { get; set; }
+    public string Ten { get; set; }
     public int TrangThai { get; set; }
 }
