+++ Shop.ViewModels/ViewModels/KichCoVM.cs
@@ -1,9 +1,9 @@
-﻿namespace Shop.Application.ViewModels;
+﻿namespace Shop.ViewModels.ViewModels;
 
-public class SanPhamVM
+public class KichCoVM
 {
     public Guid Id { get; set; }
     public string Ma { get; set; }
-    public string Ten { get; set; }
+    public int SoSize { get; set; }
     public int TrangThai { get; set; }
 }
