+++ Team5_ConverseStore.sln
@@ -11,11 +11,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shop.WebApp", "Shop.WebApp\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shop.Application", "Shop.Application\Shop.Application.csproj", "{BACACB95-7877-4A97-92F7-1C4AE6FB368C}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1.Domain", "1.Domain", "{4A1CB931-7DFD-4B96-BA47-F02D9F03A4E6}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2.Domain", "2.Domain", "{4A1CB931-7DFD-4B96-BA47-F02D9F03A4E6}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2.Hosts", "2.Hosts", "{D6774886-D49B-4C95-AAA7-658BA400F300}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "3.Hosts", "3.Hosts", "{D6774886-D49B-4C95-AAA7-658BA400F300}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shop.AdminApp", "Shop.AdminApp\Shop.AdminApp.csproj", "{26E2D827-139F-477C-8375-3AC700F14BBC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shop.AdminApp", "Shop.AdminApp\Shop.AdminApp.csproj", "{26E2D827-139F-477C-8375-3AC700F14BBC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1.Infrastructure", "1.Infrastructure", "{2806B082-8797-4149-921A-885CA5C528AF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shop.ViewModels", "Shop.ViewModels\Shop.ViewModels.csproj", "{54D948B3-DEF1-4A20-8264-7BABD2B62586}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,6 +47,10 @@ Global
 		{26E2D827-139F-477C-8375-3AC700F14BBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26E2D827-139F-477C-8375-3AC700F14BBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26E2D827-139F-477C-8375-3AC700F14BBC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{54D948B3-DEF1-4A20-8264-7BABD2B62586}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54D948B3-DEF1-4A20-8264-7BABD2B62586}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54D948B3-DEF1-4A20-8264-7BABD2B62586}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54D948B3-DEF1-4A20-8264-7BABD2B62586}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -53,6 +61,7 @@ Global
 		{94E0ED11-4B00-4058-B61B-E863F352350C} = {D6774886-D49B-4C95-AAA7-658BA400F300}
 		{BACACB95-7877-4A97-92F7-1C4AE6FB368C} = {4A1CB931-7DFD-4B96-BA47-F02D9F03A4E6}
 		{26E2D827-139F-477C-8375-3AC700F14BBC} = {D6774886-D49B-4C95-AAA7-658BA400F300}
+		{54D948B3-DEF1-4A20-8264-7BABD2B62586} = {2806B082-8797-4149-921A-885CA5C528AF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {91FE457A-4D32-474C-A2B7-79B09F9F016B}
